Refactor: retirada de logica pesada do app.py
